+++ CHANGELOG.md
@@ -37,7 +37,14 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
-None
+### Features
+
+* Marker governance proposal are supported in cli [#367](https://github.com/provenance-io/provenance/issues/367)
+
+### Bug Fixes
+
+* Marker Withdraw Escrow Proposal type is properly registered [#367](https://github.com/provenance-io/provenance/issues/367)
+  * Target Address field spelling error corrected in Withdraw Escrow and Increase Supply Governance Proposals.
 
 ## [v1.5.0](https://github.com/provenance-io/provenance/releases/tag/v1.5.0) - 2021-06-23